moved JS.util into separate file
