+++ package.json
@@ -31,6 +31,8 @@
     "test": "test"
   },
   "scripts": {
-    "test": "mocha"
+    "test": "mocha",
+    "lint": "eslint .",
+    "lint-fix": "eslint --fix ."
   }
 }
